+++ src/components/canvas/PotoCanvasClipGrid.tsx
