+++ src/components/canvas/PotoCanvasClipper.tsx
